# api-gateway/docs/swagger.yaml
definitions:
...
  orderproto.CreateOrderReq:
    properties:
      address:
        type: string
      product:
        items:
          $ref: '#/definitions/orderproto.ProductRes'
        type: array
      user_id:
        type: integer
    type: object
  orderproto.CreateOrderRes:
    properties:
      order_id:
        type: string
    type: object
  orderproto.GetOrderRes:
    properties:
      address:
        type: string
      order_id:
        type: string
      product:
        items:
          $ref: '#/definitions/orderproto.ProductRes'
        type: array
      status:
        type: string
      totalamount:
        type: number
      user_id:
        type: integer
    type: object
  orderproto.ProductRes:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
    type: object
  orderproto.UpdateOrderRes:
    properties:
      message:
        type: string
    type: object
  orderproto.UpdateReq:
    properties:
      address:
        type: string
      order_id:
        type: string
      product:
        items:
          $ref: '#/definitions/orderproto.ProductRes'
        type: array
    type: object
  paymentproto.ProcessPaymentRequest:
    properties:
      order_id:
        type: string
      payment_details:
        type: string
      payment_method:
        type: string
      user_id:
        type: integer
    type: object
  paymentproto.ProcessPaymentResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
      transaction_id:
        type: string
    type: object
  paymentproto.RefundPaymentRequest:
    properties:
      reason:
        type: string
      refund_amount:
        type: number
      transaction_id:
        type: string
    type: object
  paymentproto.RefundPaymentResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  productproto.CreateReq:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  productproto.CreateRes:
    properties:
      message:
        type: string
    type: object
  productproto.ListProduct:
    properties:
      product:
        items:
          $ref: '#/definitions/productproto.Product'
        type: array
    type: object
  productproto.Product:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
...
info:
  contact: {}
  description: This is a sample server for a restaurant reservation system.
  title: Food Delivery
  version: "1.0"
paths:
...
  /orders:
    post:
      consumes:
      - application/json
      description: Creates a new order with the provided details
      parameters:
      - description: Order request body
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/orderproto.CreateOrderReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/orderproto.CreateOrderRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer Auth: []
      summary: Create a new order
      tags:
      - order
    put:
      consumes:
      - application/json
      description: Updates an existing order with the provided details
      parameters:
      - description: Order update request
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/orderproto.UpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/orderproto.UpdateOrderRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer Auth: []
      summary: Update an order
      tags:
      - order
  /orders/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an order by its ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/orderproto.UpdateOrderRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer Auth: []
      summary: Delete an order
      tags:
      - order
    get:
      consumes:
      - application/json
      description: Retrieves an order by its ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/orderproto.GetOrderRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer Auth: []
      summary: Get order by ID
      tags:
      - order
  /payment/process:
    post:
      consumes:
      - application/json
      description: Processes a payment based on the provided payment details
      parameters:
      - description: Payment processing request
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/paymentproto.ProcessPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/paymentproto.ProcessPaymentResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer Auth: []
      summary: Process a payment
      tags:
      - payment
  /payment/refund:
    post:
      consumes:
      - application/json
      description: Refunds a payment based on the provided refund details
      parameters:
      - description: Payment refund request
        in: body
        name: refund
        required: true
        schema:
          $ref: '#/definitions/paymentproto.RefundPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/paymentproto.RefundPaymentResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer Auth: []
      summary: Refund a payment
      tags:
      - payment
  /products:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/productproto.ListProduct'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer Auth: []
      summary: Get all products
      tags:
      - product
    post:
      consumes:
      - application/json
      description: Create a new product with the given details
      parameters:
      - description: Product creation request
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/productproto.CreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/productproto.CreateRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer Auth: []
      summary: Create a new product
      tags:
      - product
    put:
      consumes:
      - application/json
      description: Update the details of an existing product
      parameters:
      - description: Product update request
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/productproto.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/productproto.CreateRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer Auth: []
      summary: Update product details
      tags:
      - product
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing product by product ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/productproto.CreateRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer Auth: []
      summary: Delete product by ID
      tags:
      - product
    get:
      consumes:
      - application/json
      description: Retrieve product details by product ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/productproto.Product'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer Auth: []
      summary: Get product by ID
      tags:
      - product
...
securityDefinitions:
  Bearer:
    description: Enter the token in the format `Bearer {token}`
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"